Move Linux FFI window code into own module
